Fix defense export (DefenseStat struct → int) + guard site against non-primitive stats
Live test with a real player revealed statDefense is a DefenseStat struct since
tML 1.4.4, so it serialized as an object; cast to int for the effective value.
Site also now skips any non-primitive stat value defensively.

+++ character-stats-mod/StatusExport.cs
@@ -113,7 +113,10 @@ namespace StatusExport
 				lifeMax = p.statLifeMax2,
 				mana = p.statMana,
 				manaMax = p.statManaMax2,
-				defense = p.statDefense,
+				// statDefense is a DefenseStat struct since tML 1.4.4 — coerce to the
+				// effective int via its implicit conversion (serializing it raw dumps
+				// {Positive,Negative,AdditiveBonus,FinalMultiplier} instead of a number).
+				defense = (int)p.statDefense,
 				difficulty,
 				inventoryCount = invCount,
 				equips,
